fixed split_tags function and move all helper functions to the utils.py
